+++ includes/class-list-table.php
@@ -91,7 +91,7 @@ class Code_Snippets_List_Table extends WP_List_Table {
 			return;
 
 		/* Load a different stylesheet if MP6 is active */
-		if ( 'mp6' === get_user_option( 'admin_color' ) ) {
+		if ( defined( 'MP6' ) ) {
 
 			wp_enqueue_style(
 				'snippets-table-mp6',
@@ -199,10 +199,16 @@ class Code_Snippets_List_Table extends WP_List_Table {
 		else
 			$title = sprintf ( __( 'Untitled #%d', 'code-snippets' ), $snippet->id );
 
+		$row_actions = $this->row_actions( $actions,
+			apply_filters( 'code_snippets/list_table/row_actions_always_visiable', false )
+		);
+
 		/* Return the name contents */
 		return apply_filters(
 			'code_snippets/list_table/column_name',
-			sprintf ( '<strong>%s</strong>', $title ) . $this->row_actions( $actions, true ),
+			sprintf ( '<a href="%2$s"><strong>%1$s</strong></a>', $title,
+				add_query_arg( 'edit', $snippet->id, $code_snippets->admin_single_url )
+			) . $row_actions,
 			$snippet
 		);
 	}